+++ infra/main.tf
@@ -675,6 +675,14 @@ resource "aws_apigatewayv2_route" "create_vehicle" {
   authorizer_id      = aws_apigatewayv2_authorizer.jwt.id
 }
 
+resource "aws_apigatewayv2_route" "delete_vehicle" {
+  api_id             = aws_apigatewayv2_api.vwc_api.id
+  route_key          = "DELETE /vehicles/{vehicleId}"
+  target             = "integrations/${aws_apigatewayv2_integration.vwc.id}"
+  authorization_type = "JWT"
+  authorizer_id      = aws_apigatewayv2_authorizer.jwt.id
+}
+
 resource "aws_apigatewayv2_route" "get_vehicle_overview" {
   api_id             = aws_apigatewayv2_api.vwc_api.id
   route_key          = "GET /vehicles/{vehicleId}/overview"